Update and rename car_lidar_pointcloud.py to sensorframe_lidar_pointcloud.py
